requirements.txt
# Core dependencies
gradio>=4.0.0
groq>=0.4.0
python-dotenv>=1.0.0

# Speech and Audio processing
speechrecognition>=3.10.0
pyaudio>=0.2.11
pydub>=0.25.1

# Text-to-Speech
gtts>=2.3.0
pygame>=2.5.0

# Image processing
pillow>=9.0.0

# Additional utilities
requests>=2.28.0
